+++ TombEngine/Objects/TR5/Entity/tr5_lion.cpp
@@ -98,7 +98,7 @@ namespace TEN::Entities::Creatures::TR5
 			item->HitPoints = 0;
 
 			if (item->Animation.ActiveState != LION_STATE_DEATH)
-				SetAnimation(item, LionDeathAnims[Random::GenerateInt(0, LionDeathAnims.size() - 1)]);
+				SetAnimation(item, LionDeathAnims[Random::GenerateInt(0, (int)LionDeathAnims.size() - 1)]);
 		}
 		else
 		{
